Decouple PlanetModal from redux store and create App.tsx to handle global logic
