+++ src/components/PlanetModal/index.tsx
@@ -3,78 +3,38 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Button, Modal, ModalBody, ModalFooter, ModalHeader } from 'reactstrap';
 import { z } from 'zod';
 
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { modalActions } from '../../slices/modal';
+import { FormData } from '../../slices/modal';
 import Input from '../Input';
 import Select from '../Select';
+import { PLANET_FORM_SCHEMA, TERRAINS } from './constants';
 
-const formSchema = z.object({
-  name: z
-    .string({
-      required_error: 'Name is required',
-    })
-    .min(1),
-  rotation_period: z.number({
-    required_error: 'Rotation Period is required',
-    invalid_type_error: 'Rotation Period must be a number',
-  }),
-  orbital_period: z.number({
-    required_error: 'Orbital Period is required',
-    invalid_type_error: 'Orbital Period must be a number',
-  }),
-  diameter: z.number({
-    required_error: 'Diameter is required',
-    invalid_type_error: 'Diameter must be a number',
-  }),
-  climate: z
-    .string({
-      required_error: 'Climate is required',
-    })
-    .min(1),
-  gravity: z
-    .string({
-      required_error: 'Gravity is required',
-    })
-    .min(1),
-  terrain: z
-    .string({
-      required_error: 'Terrain is required',
-    })
-    .min(1),
-  surface_water: z.number({
-    required_error: 'Surface Water is required',
-    invalid_type_error: 'Surface Water must be a number',
-  }),
-});
+type Schema = z.infer<typeof PLANET_FORM_SCHEMA>;
 
-type Schema = z.infer<typeof formSchema>;
+interface PlanetModalProps {
+  defaultValues: FormData;
+  onClose: () => void;
+  onSubmit: (data: Schema) => void;
+}
 
-const TERRAINS =
-  'desert,grasslands,mountains,tundra,jungle,ice caves,swamp'.split(',');
-
-const PlanetModal = () => {
-  const dispatch = useAppDispatch();
-  const defaultValues = useAppSelector((state) => state.modal.formData);
+const PlanetModal = (props: PlanetModalProps) => {
+  const { defaultValues, onClose, onSubmit } = props;
 
   const formMethods = useForm<Schema>({
     defaultValues,
     mode: 'onChange',
-    resolver: zodResolver(formSchema),
+    resolver: zodResolver(PLANET_FORM_SCHEMA),
   });
 
-  const closeModal = () => dispatch(modalActions.clearData());
-  const submitModal = (data: Schema) => dispatch(modalActions.submitData(data));
-
   return (
-    <Modal isOpen centered backdrop toggle={closeModal}>
+    <Modal isOpen centered backdrop toggle={onClose}>
       <ModalHeader>Edit Planet {defaultValues.name}</ModalHeader>
 
       <ModalBody>
         <FormProvider {...formMethods}>
           <form
-            id="planet-form"
+            id="planet_form"
             className="d-flex flex-column"
-            onSubmit={formMethods.handleSubmit(submitModal)}
+            onSubmit={formMethods.handleSubmit(onSubmit)}
           >
             <Input<Schema> name="name" label="Planet Name" required />
 
@@ -134,11 +94,11 @@ const PlanetModal = () => {
       </ModalBody>
 
       <ModalFooter>
-        <Button color="danger" onClick={closeModal}>
+        <Button color="danger" onClick={onClose}>
           Cancel
         </Button>
         <Button
-          form="planet-form"
+          form="planet_form"
           type="submit"
           color="primary"
           disabled={!formMethods.formState.isValid}
